Keep AXContexts alive for child frames when a11y domain is enabled
Before we would only keep the context associated with the root frame
alive. With this change, we maintain a map from `LocalFrame` to
`AXContext` such that object caches are kept for both the root document
and child documents when the accessibility domain is enabled.

Bug: 1226486
Change-Id: I7fe7b619b94b70b5b1a4669f0086c639ea0a65e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3070918
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Johan Bay <jobay@chromium.org>
Cr-Commit-Position: refs/heads/master@{#909723}