Reset a flag on first update event in a gesture
The flag not being reset to false means that if the last update event in
a previous gesture was marked as janky, the flag for "previous event
reported" would be set to true. This means that the first event in the
next gesture would never be able to be reported, which has been observed
in recorded traces while investigating discrepancy between UMA and
trace-based jank metrics.

This CL extracts all the per-scroll state into a struct, allowing to
reset all the variables in one line, avoiding this and potentially
other similar problems.

Bug: b/185991751
Change-Id: I178ef4caa2e224445d83521160c34ebeec069b1b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3077945
Commit-Queue: Andrew Shulaev <ddrone@google.com>
Reviewed-by: Stephen Nusko <nuskos@chromium.org>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Cr-Commit-Position: refs/heads/master@{#909767}