fast_ink: Match color formats between GMB and SharedImage
For Fast Ink, Chrome creates the backing buffer with a format
dependent on SK_B32_SHIFT, either RGBA or BGRA [1]. When we create a
SharedImage from the GMB, however, the wrapping TransferableResource
always claims a RGBA format [2]. I don’t think this is too big of a
deal in practice, but when running ash_unittests with SkiaRenderer
enabled, a number of ink tests fail a format DCHECK [3] in the Skia
API. Here, |surface->backendFormat()| is populated by the GMB format
while |fFormat| comes from the TransferableResource format.

This CL fixes the issue by also applying a conditional on
SK_B32_SHIFT when creating the SharedImage so the formats match.

[1] https://source.chromium.org/chromium/chromium/src/+/main:ash/fast_ink/fast_ink_host.cc;l=256;drc=54c89f851b4160468874d8903663dba85e99101b
[2] https://source.chromium.org/chromium/chromium/src/+/main:ash/fast_ink/fast_ink_host.cc;l=385;drc=54c89f851b4160468874d8903663dba85e99101b
[3] https://source.chromium.org/chromium/chromium/src/+/main:third_party/skia/src/gpu/GrSurfaceProxy.cpp;l=466;drc=f3868628f987cb42b0a95cd62b9867ded4504523

Bug: b:195938645
Change-Id: Ie3343fb6c4c5d36c9348a03791b39537fbb7a82a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3081761
Commit-Queue: Brian Ho <hob@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Cr-Commit-Position: refs/heads/master@{#909816}