[Shelf] Wrap the bounds of the dragged shelf item in absl::optional
We should wrap `ScrollableShelfView::drag_item_bounds_in_screen_` in
absl::optional and reset it when shelf drag-and-drop ends. It helps to
indicate whether shelf drag-and-drop is ongoing.

This bug is exposed by a recent CL which refactors shelf drag-and-drop
(https://crrev.com/c/3046704).

Bug: 1235882
Change-Id: I0ace13c51164d6eed72bf6777f579df74fd5cc59
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3079060
Reviewed-by: Toni Baržić <tbarzic@chromium.org>
Commit-Queue: Andrew Xu <andrewxu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#909817}