Address an issue breaking the FSC tests.
A DownloadItemImpl will create its RenameHandler only once, and cache
the object.

At creation time, RenameHandler requires that its associated
DownloadItem has the full path set.

So if the DownloadItemImpl creates the RenameHandler before the full
path is set - that is, before download completes, then the RenameHandler is
unusable.

But that is what is happening in the FSC tests: the FSC test'
BoxDownloadItemObserver receives updates on the DownloadItem. At each
update, the observer queries the DownloadItem to see if it has a
RenameHandler. This query triggers the DownloadItemImpl to create the
RenameHandler, before it is appropriate to create one.

Bug: 1214825
Change-Id: I0b2c8a2c5143bf8e5fd44690b7ff220cd0daaf58
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3078752
Reviewed-by: Alice Gong <alicego@google.com>
Commit-Queue: Alice Gong <alicego@google.com>
Cr-Commit-Position: refs/heads/master@{#909820}