Allow no-op stretch filters to be constructed on non-android.
In https://crbug.com/1226170 the stretch shader was compiled out of
non-android platforms. A NOTREACHED() was added as a safety to ensure
that we did not try to use the no-op stretch shader on these platforms,
however because shaders are passed between processes all values must be
valid, even if they aren't meant to be used. This allows the shader to
be used as effectively a no-op.

Bug: 1237642
Change-Id: I7186177fcd48ae75c31ae2521104825265debee7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3081962
Auto-Submit: Robert Flack <flackr@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#909854}