History WebUI: Update roles of date heading
Screen readers seem to ignore elements that are not in a |row| or
|gridcell|, so this CL separates out the data header into its own row
and uses the |rowheader| role to denote that the header is a row header.

The |heading| role is kept onto the heading itself so that screen
readers can use heading-related shortcuts to navigate to it and for the
screen reader to read that the heading is level 2.

Bug: 1187212
Change-Id: Ie67006c6f924888307bb0b4165ec8eddb3012194
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3077788
Reviewed-by: Rebekah Potter <rbpotter@chromium.org>
Commit-Queue: John Lee <johntlee@chromium.org>
Cr-Commit-Position: refs/heads/master@{#909917}