[ntp][modules][taskV2] Reland "Make new recipe Task module"
This is the reland for making the new recipe task module.
This is the second version of the Task module for recipes, implemented
according to the V2 mocks. We are creating a separate module to
implement V2 to ensure the original version of the module continues
to work and remains unchanged. (This is not a complete module. It's
a frame for a module, and we will add content to it in future CLs.)
When the modules-redesigned-enabled flag is set, the new V2 recipes
Task module will be displayed. Otherwise, the original version will be
displayed.

* Create a new (recipes) Task module.
* When modules redesigned flag is enabled,
  display this new Task module.
* Add unit test to test if module appears on render.

Bug: 1237495
Change-Id: I515be3fd1b598ff9002175fdc982c112af3d7c8b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3078441
Reviewed-by: Tibor Goldschwendt <tiborg@chromium.org>
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Commit-Queue: Deanna Oei <doei@google.com>
Cr-Commit-Position: refs/heads/master@{#909942}