settings: Deeplinking app notifications subpage and DND toggle
This CL adds the search results for the app notifications subpage
itself and the Do Not Disturb toggle depending on whether it is
currently On or Off. A new method, IsQuietMode(), was also added to
retrieve the current state of the quiet mode and to determine the
appropriate search result for the toggle.

Bug: 1217174
Change-Id: I308c115ea926f1d92840f0639a1309863e3807d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3061641
Reviewed-by: Jimmy Gong <jimmyxgong@chromium.org>
Reviewed-by: James Cook <jamescook@chromium.org>
Reviewed-by: Regan Hsu <hsuregan@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Caroline Hana <carolinehana@google.com>
Cr-Commit-Position: refs/heads/master@{#909950}