Lookalikes: Implement character swap heuristic for top domains
This CL uses the character swap heuristic to detect sites that are
lookalikes of top 500 sites. For example, if google[.]com is a top site
and the user navigates to goolge[.]com, the heuristic detects this as
a lookalike match.

The heuristic currently doesn't show any UI, it simply logs the match
and continues.

Bug: 1109056
Change-Id: I43a2faf66186c12098a51f164b79b20d2b0d5ea5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3070140
Commit-Queue: Mustafa Emre Acer <meacer@chromium.org>
Reviewed-by: Joe DeBlasio <jdeblasio@chromium.org>
Cr-Commit-Position: refs/heads/master@{#909961}