Set overscroll mode only when recycler view is created
We need to disable overscroll mode for recycler view when start
surface is enabled. This is because RecyclerView does not manage
all the start surface page content and we don't want the overscroll
effect to appear just for RecyclerView. Previous fix for this
doesn't take into account that RecyclerView may not be created in
certain circumstance.

Bug: 1237238
Change-Id: I2b217edf697a077fb1ae620ce9d04d2520ca4137
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3078301
Commit-Queue: Jian Li <jianli@chromium.org>
Reviewed-by: Xi Han <hanxi@chromium.org>
Reviewed-by: Sky Malice <skym@chromium.org>
Cr-Commit-Position: refs/heads/master@{#909991}