Remove InstallableInkDrop
InstallableInkDrop is no longer a requirement to install an InkDrop on
any View, and the feature flag for using it has expired long ago without
any current plans to reinstate it. If we do end up wanting to pick this
up again for a subset of Views to use as their InkDrop implementation,
then reverting this change is still an option.

Bug: 931964, 1207526
Change-Id: Ibb57da58079b6609dac0624b7e7f4272898a2527
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3077527
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Peter Boström <pbos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#910006}