Add Android setting for HTTPS-First Mode
This adds a new setting to the "Privacy and Security" settings subpage
for HTTPS-First Mode, and hooks it up to the native pref and enterprise
policy.

Note that new code added in this CL refers to this as "HTTPS-First Mode"
but still references or calls into old code that still refers to this as
"HTTPS-Only Mode". A followup CL will handle more of the renaming.

Bug: 1218526
Change-Id: I25cba0f6323b81648e14428e5622a2f6e627aaa1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3072967
Reviewed-by: Christian Dullweber <dullweber@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Reviewed-by: Carlos IL <carlosil@chromium.org>
Commit-Queue: Chris Thompson <cthomp@chromium.org>
Cr-Commit-Position: refs/heads/master@{#910011}