Add MultiProfileDownloadNotifier abstraction
MultiProfileDownloadNotifier wraps multiple AllDownloadItemNotifiers and
handles the responsibility of creating and destroying those notifiers as
their DownloadManagers' Profiles are created and destroyed. Notably, a
MultiProfileDownloadNotifier has the capability of recognizing when a
profile that it observes spawns a child profile, and it will observe the
downloads from that profile as well.

Bug: 1208910
Change-Id: I2cdf6511b0aa8ee5f12b76b48f49735df07bbf70
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3068044
Commit-Queue: Colin Kincaid <ckincaid@chromium.org>
Reviewed-by: Min Qin <qinmin@chromium.org>
Reviewed-by: Xing Liu <xingliu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#910027}