settings: Fix mojo receiver binding for App Notifications Subpage
Previously, when reloading the App Notification Subpage several times
the settings page would crash due to the receiver already bounded.

This Cl resolves this issue by adding a check to BindInterface() in the
App Notification handler that resets and re-binds the receiver if it is
currently bound.

Bug: 1217174
Change-Id: Iac77e474cf7068c1dc3cf96d420b28d5d434dc86
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3078584
Commit-Queue: Caroline Hana <carolinehana@google.com>
Reviewed-by: Jimmy Gong <jimmyxgong@chromium.org>
Reviewed-by: Regan Hsu <hsuregan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#910048}