Load more feed content if there are no off-screen-scrollable cards.
The Chrome Discover feature attempts to load more content when the
user scrolls such the last displayed card is fewer than the "load
more threshold" away from the end of the feed.

This CL addresses a corner case in which, upon load, all of the
feed cards are already displayed on the screen. In this scenario,
the feed may not be scrollable, which leaves the user unable to
initiate loading more cards.

When the feed content is updated (i.e., loaded) we now check if
there are zero off-screen cards to which the user can scroll. If
so, we load more content.

This CL also adds UMA histograms tracking the length of the feed
and the number of not-yet-scrolled-to cards when a refresh is
triggered.

Bug: 1195148
Change-Id: Ib514dd60d7e2e837192d43aa14fee112c338ca18
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2883981
Auto-Submit: Roger McFarlane <rogerm@chromium.org>
Commit-Queue: Roger McFarlane <rogerm@chromium.org>
Reviewed-by: Cathy Li <chili@chromium.org>
Reviewed-by: Justin DeWitt <dewittj@chromium.org>
Reviewed-by: Dan H <harringtond@chromium.org>
Cr-Commit-Position: refs/heads/master@{#910055}