Convert AppServiceInstanceRegistry::OnInstances to use InstanceKey
aura::Window* will not be available in Ash for web apps that run in
tabs after Lacros is enabled. Any code that uses an aura::Window* to
refer to an instance should instead use an InstanceKey unless it is
guaranteed to only be referring to top level aura::Window*s (which
will still be available in Ash with Lacros).

This CL changes the OnInstances function to use InstanceKey.

BUG=1235206

Change-Id: Ida2aef00faf2c7da96823e34e0a3e066b58ae860
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3073481
Commit-Queue: Raymes Khoury <raymes@chromium.org>
Reviewed-by: Nancy Wang <nancylingwang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#910066}