[ntp][modules] Use mouse events instead of drag events
The modules drag and drop was originally implemented using the HTML Drag
& Drop API but by default when there is no drag image or link, a globe
icon is attached to the cursor during dragging. However, for our purpose
the globe icon is unnecessary because we stick the module itself to the
cursor, so this CL uses mouse events instead of drag events. This allows
for the same drag and drop functionality but without the globe icon.

Fixed: 1238041
Change-Id: I7e02b16d049f389a56d51a8157c6e0cc5e8245e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3082216
Commit-Queue: Lydia Lam <lydialam@google.com>
Reviewed-by: Tibor Goldschwendt <tiborg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#910121}