[CrOS Device Name] Make DeviceNameStore an observer of UserManager
This CL adds DeviceNameStore class as an observer of UserManager class
because user profile is loaded after sign-in, which makes the user a
device owner. This should change device name state and in turn notify
observers of DeviceNameStore. Unit tests are added to test this scenario

Bug: 126802
Change-Id: I065523ee5f550c27625225af13b44877505d185d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3083982
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Commit-Queue: Jason Ah Chuen <ahchuen@google.com>
Cr-Commit-Position: refs/heads/master@{#910167}