MPArch: Don't send TitleWasSet for prerender entry title updates
Currently WebContentsImpl::UpdateTitle sets a title for the
NavigationEntry of a prerendering page and notifies observers that a
title was set for the entry. Exposing a prerendering entry to observers
is something we want to avoid.

This CL stops TitleWasSet being sent while a page is prerendering, and
only sends it after the page is activated.

Change-Id: I5fdaf016948e71200eaa78a0ec676fa3e431899f
Bug: 1208556
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2983059
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Reviewed-by: Lucas Gadani <lfg@chromium.org>
Commit-Queue: Adithya Srinivasan <adithyas@chromium.org>
Cr-Commit-Position: refs/heads/master@{#910314}