tests: Increase next frame timeout for OverviewSessionTest.*
Apologies, but the investigation into this issue was a bit
complicated, and there's a fair amount of explanation to justify
this change. Please refer to the bug for more information:
https://b.corp.google.com/issues/195756769

This CL increases the next frame timeout in the OverviewSessionTest
suite to account for the ~300ms frame resulting from SwiftShader's
slow draw call. This fixes the test suite when running on
SkiaRenderer and should also deflake [1] the tests on GLRenderer
since in my local testing, even GLRenderer sometimes produced frames
close to the 100ms deadline (e.g. 80ms).

[1] https://analysis.chromium.org/p/chromium/flake-portal/flakes/occurrences?key=ag9zfmZpbmRpdC1mb3ItbWVyRgsSBUZsYWtlIjtjaHJvbWl1bUBhc2hfdW5pdHRlc3RzQE92ZXJ2aWV3U2Vzc2lvblRlc3QuRnVsbHNjcmVlbldpbmRvdww

Bug=b:195938645, 1145226

Change-Id: Id3dee1092622338efbc45e58b98a28f012f9a454
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3076624
Commit-Queue: Brian Ho <hob@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Cr-Commit-Position: refs/heads/master@{#910326}