Offline Dictation: Show notification when SODA download finishes
This change adds behavior to show a system notification when SODA
download finishes. Note, we only do this if Dictation was triggered by
the user. Please see the bug for a link to the UX mock and other
details.

This behavior is hidden behind the offline Dictation feature, which
can be enabled by using the command line parameter:

--enable-features="ExperimentalAccessibilityDictationOffline"

Bug: 1234123
Change-Id: Iaa1c577d7cb57bf7de48b6dcbf44e4f4ea988e4f
AX-Relnotes: N/A
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3059501
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Abigail Klein <abigailbklein@google.com>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Justin DeWitt <dewittj@chromium.org>
Reviewed-by: Katie Dektar <katie@chromium.org>
Commit-Queue: Akihiro Ota <akihiroota@chromium.org>
Cr-Commit-Position: refs/heads/master@{#910327}