[attribution_reporting] Retry failed reports without response headers
What:
Adds a layer of retry logic where reports will be re-queued if they fail
to send with an internal network error, beyond the current initial
retry.

Requests are only retried for a subset of network errors which have
been seen to be problematic.

Why:
10% of reports fail to send due to internal network errors, see:

https://groups.google.com/a/chromium.org/g/attribution-reporting-api-dev/c/2KKqWG70w7U/m/7M8fWEbmAgAJ

How:
Do not delete conversion reports from storage whose report fails
to send with a net:Error.

Add logic which only deletes on successful send, or a report is
sufficiently old. Reports are expired after 14 days, chosen from
looking at the Conversions.ExtraReportDelay histogram.

Reports that are re-queued will need noise added to them so that they
are not temporally joinable.

This change also re-organizes tests in ConversionNetworkSenderImpl
to be grouped in order of complexity + similar behavior.

Bug: 1054127
Change-Id: Ic320cdf13ce590369eaed9994d9cecbd24549275
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3028422
Commit-Queue: John Delaney <johnidel@chromium.org>
Reviewed-by: Andrew Paseltiner <apaseltiner@chromium.org>
Cr-Commit-Position: refs/heads/master@{#910336}