[ozone] Hidden "Use system title bar" when it is not supported.
On Linux, it is possible to have multiple graphics environments with
different capabilities on the same host.  For instance, at the moment of
writing, KDE Plasma exposes zxdg_decoration_manager while GNOME does
not.  This makes possible the situation when the user first enables
server-side decorations in the graphics session that allows that, and
after changing to another session that does not support server-side
decorations, Chromium launches without any decorations because the user
prefers to use server-side ones, but these are not available, and as the
settings UI for enabling them is hidden, the user cannot fix that.

This CL fixes that by 1) forcing the custom frame (a.k.a. client-side
decorations) when they are suggested by the platform and 2) hiding the
menu item that toggles the decoration type when the system does not
support server-side ones.

Bug: 1138724
Change-Id: If6b6b2f35c63811789e15cc3f4513d9747f7bbbe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3075386
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Alexander Dunaev <adunaev@igalia.com>
Cr-Commit-Position: refs/heads/master@{#910360}