Remove default toolbar from BottomSheet
The default toolbar is a long-obsolete view that was used to fill in
the background of the toolbar section of the sheet while there was
no content to be shown. The sheet now has a background that handles
this case. No other logic depends on necessarily having a toolbar be
present.

Bug: 986310
Change-Id: Iaad5ea628c48a384f110e12c954d76ae833848bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3082489
Reviewed-by: Sinan Sahin <sinansahin@google.com>
Commit-Queue: Matt Jones <mdjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#910371}