content: Remove RenderFrameHostImpl::AccessibilityWebContents.
This is a layering violation as RenderFrameHostImpl is in renderer_host/
and should not depend on WebContents. The current code bypassed the
DEPS check by using a forward declaration via
render_frame_host_delegate.h which exposes GetAsWebContents(). This
CL removes one usage of GetAsWebContents() to help remove the function,
as tracked in the linked bug.

Bug: 82582
Change-Id: I396e1ea86895ae3c12792475cba998708d6720ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3077911
Commit-Queue: Matt Falkenhagen <falken@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#910372}