[omnibox] Add a is_active column to keyword table
This CL adds an is_active column to the keyword table. This value will
be used to turn on/off search engines so we can disable auto-added
engines by default without disrupting users as much as possible.

It is currently not synced and not hooked up to anything.

Design here: go/scoped-search-technical-design

Bug: 1213659
Change-Id: I04a0d98c19a54a666bc4a7b082b6ca135436dea4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3076530
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Commit-Queue: Angela Yoeurng <yoangela@chromium.org>
Cr-Commit-Position: refs/heads/master@{#910434}