ConversionStorageSql init status histogram doesn't support all values
Currently, it uses a form of base::UmaHistogramEnumeration that expects
an enum_size as the last parameter, but it instead provides a max value
(likely due to confusing documentation). This change means that trying
to record kFailedToInitializeSchema to the histogram will *not* hit a
DCHECK. Given this change, we also version the histogram.

Change-Id: Ia48c5379ca0dbf67ee5ffb2fe6942c2ce4e8e820
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3072540
Reviewed-by: John Delaney <johnidel@chromium.org>
Commit-Queue: Alex Turner <alexmt@chromium.org>
Cr-Commit-Position: refs/heads/master@{#910457}